Further improve speed.
The "phase" variable is unnecessary and the pump pins can be cycled by
simply XORing PORTB with the togglePort variable.

By eliminating the phase variable and directly working with the PORTB
register, pumping speed is increased to 450kHz with an 8MHz system
clock.

Also, move the initialization state for PORTB (i.e., pin P1 is high) to
setup() instead of having it in loop().